+++ packages/foundations/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.1.4
+
+- [#719](https://github.com/iTwin/design-system/pull/719): Updated `Icon` component to catch errors when making network requests.
+- [#650](https://github.com/iTwin/design-system/pull/650): Added global `::selection` styling.
+
+## 0.1.3
+
+Updated internal code for `@stratakit/bricks@0.2.0` and `@stratakit/structures@0.1.1`.
+
 ## 0.1.2
 
 Updated styling for `@stratakit/bricks@0.1.2`.